requirements.txt
aiohttp==3.10.5
aiolimiter==1.1.0
SQLAlchemy==2.0.34
asyncpg==0.29.0
alembic==1.13.2
python-dotenv==1.0.1
tenacity==9.0.0
structlog==24.2.0
orjson==3.10.7
